torch.arange(8, dtype=torch.uint64, device="cuda")

+++ c/chessintionlib/chess_aux_c.py
@@ -77,7 +77,7 @@ def concat_fen_legal_bits(fen):
     compressed_tensor = torch.from_numpy(compressed_array).to(dtype=torch.uint64, device="cuda")
 
     # Expande los bits con una operación eficiente
-    bit_tensor = torch.bitwise_and(torch.bitwise_right_shift(compressed_tensor[:, None], torch.arange(8, device="cuda")), 1).bool()
+    bit_tensor = torch.bitwise_and(torch.bitwise_right_shift(compressed_tensor[:, None], torch.arange(8, dtype=torch.uint64, device="cuda")), 1).bool()
 
     # Darle forma (77, 8, 8)
     return bit_tensor.view(77, 8, 8)
